// src/app/interfaces/UserInterface.ts
export interface UserInterface {
    id: number;
    username: string;
    password: string;
    firstname: string;
    lastname: string;
    email: string;
    amountOfPeopleInTheResidence: number;
    firstAccess: number;
    enabled: boolean;
    cpf:string;
    token:string;
    adress: string;
 }